backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Redis
redis==5.2.0

# AI/ML APIs
openai==1.54.0
google-generativeai==0.8.3
groq==0.11.0

# Vector DB
chromadb==0.5.20
sentence-transformers==3.3.1

# Audio Processing (Google Cloud TTS - optional)
google-cloud-texttospeech==2.18.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP & Utils
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Monitoring
loguru==0.7.2
